refactor: [FE] checkIsNotGuidePage를 훅으로 분리
