+++ frontend/src/components/QuizResultContent.tsx
@@ -1,17 +1,13 @@
 import styled from 'styled-components';
 import { Center } from '@styles/styled';
 import { useRecoilValue } from 'recoil';
-import { currentPageIdxState, currentSketchbookState, isEndedState } from '@atoms/result';
+import { currentSketchbookState } from '@atoms/result';
+import useCheckGuidePage from '@hooks/useCheckGuidePage';
 
 function QuizResultContent() {
-    const currentPageIdx = useRecoilValue(currentPageIdxState);
-    const isEnded = useRecoilValue(isEndedState);
     const currentSketchbook = useRecoilValue(currentSketchbookState);
-    const guidePageIdx = -1;
+    const { checkIsNotGuidePage } = useCheckGuidePage();
 
-    function checkIsNotGuidePage() {
-        return currentPageIdx > guidePageIdx && !isEnded;
-    }
     return (
         <>
             {checkIsNotGuidePage() && (
